cast sqlite3_column_count result to NSUInteger
the `sqlite3_column_count` return an int but the `initWithCapacity` need
an `NSUInteger`, i.e. the result need a type cast.

+++ RASqlite/RASqlite.m
@@ -439,7 +439,7 @@ static NSString *RASqliteNestedTransactionException = @"Nested transactions";
 }
 
 - (NSDictionary *)fetchColumns:(sqlite3_stmt **)statement {
-    unsigned int count = sqlite3_column_count(*statement);
+    NSUInteger count = (NSUInteger) sqlite3_column_count(*statement);
     NSMutableDictionary *row = [[NSMutableDictionary alloc] initWithCapacity:count];
 
     const char *name;
